fix(intercom): fix auth
