+++ packages/pieces/community/intercom/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@activepieces/piece-intercom",
-  "version": "0.6.5",
+  "version": "0.7.0",
   "main": "./dist/src/index.js",
   "types": "./dist/src/index.d.ts",
   "dependencies": {
