+++ packages/pieces/community/intercom/src/index.ts
@@ -51,70 +51,8 @@ import { leadConvertedToUserTrigger } from './lib/triggers/lead-converted-to-use
 import { newUserTrigger } from './lib/triggers/new-user';
 import { tagAddedToUserTrigger } from './lib/triggers/tag-added-to-user';
 import { contactUpdatedTrigger } from './lib/triggers/contact-updated';
-
-const regionProp = Property.StaticDropdown({
-	displayName: 'Region',
-	required: true,
-	options: {
-		options: [
-			{ label: 'US', value: 'intercom' },
-			{ label: 'EU', value: 'eu.intercom' },
-			{ label: 'AU', value: 'au.intercom' },
-		],
-	},
-});
-
-const oauthDescription = `
-Please follow the instructions to create Intercom Oauth2 app.
-
-1.Log in to your Intercom account and navigate to **Settings > Integrations > Developer Hub**.
-2.Click on **Create a new app** and select the appropriate workspace.
-3.In **Authentication** section, add Redirect URL.
-4.In **Webhooks** section, select the events you want to receive.
-5.Go to the **Basic Information** section and copy the Client ID and Client Secret.
-`;
-
-export const intercomOAuth2Auth = PieceAuth.OAuth2({
-	authUrl: 'https://app.{region}.com/oauth',
-	tokenUrl: 'https://api.{region}.io/auth/eagle/token',
-	required: true,
-	description: oauthDescription,
-	scope: [],
-	props: {
-		region: regionProp,
-	},
-});
-
-const intercomCustomAuth = PieceAuth.CustomAuth({
-	displayName: 'Access Token',
-	description:
-		'Connect using an Intercom Access Token. You can find your token in **Settings > Integrations > Developer Hub > Your App > Authentication**.',
-	required: true,
-	props: {
-		accessToken: PieceAuth.SecretText({
-			displayName: 'Access Token',
-			required: true,
-		}),
-		region: regionProp,
-	},
-	validate: async ({ auth }) => {
-		try {
-			await httpClient.sendRequest({
-				method: HttpMethod.GET,
-				url: `https://api.${auth.region}.io/me`,
-				authentication: {
-					type: AuthenticationType.BEARER_TOKEN,
-					token: auth.accessToken,
-				},
-			});
-			return { valid: true };
-		} catch (e) {
-			return { valid: false, error: (e as Error).message };
-		}
-	},
-});
-
-export const intercomAuth = [intercomOAuth2Auth, intercomCustomAuth];
+import { intercomAuth } from './lib/auth';
+import { assignConversationAction } from './lib/actions/assign-conversation-to-admin';
 
 export const intercom = createPiece({
 	displayName: 'Intercom',
